Reorganize test assets
