+++ tests/reproduction/array_hard.toml
